chatbot-service-backup/requirements.txt
# JECRC RAG-Enhanced Chatbot Requirements

# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Environment Variables
python-dotenv==1.0.0

# Google Gemini API
google-generativeai==0.3.2

# RAG System Dependencies
chromadb==0.4.15
sentence-transformers==2.2.2
torch>=2.6.0
transformers>=4.35.0

# Document Processing
PyPDF2==3.0.1
python-docx==0.8.11
openpyxl==3.1.2

# Additional utilities
requests==2.31.0
jsonschema==4.19.2
numpy>=1.24.0

# Production server
gunicorn==21.2.0